+++ lib/kitchen/directions/bake_injected_exercise/bake_injected_exercise.rb
@@ -1,16 +1,15 @@
 # frozen_string_literal: true
 
 module Kitchen::Directions::BakeInjectedExercise
-  def self.v1(exercise:, alphabetical_multiparts: false, alphabetical_multiparts_block: false)
+  def self.v1(exercise:, alphabetical_multiparts: false)
     V1.new.bake(
       exercise: exercise,
-      alphabetical_multiparts: alphabetical_multiparts,
-      alphabetical_multiparts_block: alphabetical_multiparts_block
+      alphabetical_multiparts: alphabetical_multiparts
     )
   end
 
   class V1
-    def bake(exercise:, alphabetical_multiparts:, alphabetical_multiparts_block:)
+    def bake(exercise:, alphabetical_multiparts:)
       question_count = exercise.injected_questions.count
 
       context = exercise&.exercise_context
@@ -26,7 +25,6 @@ module Kitchen::Directions::BakeInjectedExercise
 
         exercise.injected_questions.each_with_index do |question, index|
           question.set(:'data-type', 'alphabetical-question-multipart')
-          question.add_class('exercise-block') if alphabetical_multiparts_block
           problem_letter = "(#{alphabet[index]})"
 
           solution = question.solution