+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Revert changes for `exercise-block` class
 * Bugfix section link replacement names (using regex)
 * Add `BakeIFrame` to Nursing External Recipe
 * Create option `problem_with_prefix` for `organic-chemistry`